Build cleanly on modern GCC/Linux x86-64
Fix the build, several 64-bit and memory bugs, and all compiler warnings.
Consolidate on the 0.95 sourcebase.

Build:
- Header-defined globals in sdlemu_overlay.h broke linking under the
  now-default -fno-common; make them extern with definitions in the .c.

64-bit correctness:
- ULONG/SLONG were unsigned/signed long, i.e. 64-bit on LP64, despite the
  "Longs should be 32-bits wide" contract the 6502/Mikie core relies on for
  address wrapping, register packing and save-state layout. Use int.
- The display and ComLynx callbacks pass a host pointer through an "objref"
  parameter typed ULONG, which truncates once ULONG is correctly 32-bit.
  Add UOBJREF (uintptr_t) and thread it through mikie, system and the SDL
  display callback.

Memory and undefined behaviour:
- overlaytype[4] received 4-char strings via strcpy, smashing the stack on
  every -format invocation.
- Four delete on new[] allocations in system.cpp.
- ContextLoad read into an int via sizeof(ULONG).
- _splitpath assigned its own Drive parameter instead of the caller's
  buffer, leaving char drive[3] uninitialized for the strcpy that follows,
  and dereferenced Filename/Extension that both callers pass as NULL.
- Two ptr = (++ptr) % N sequence-point violations in the UART RX ring.
- Uninitialized surfacewidth/surfaceheight when LynxRotate hits no case,
  stale "divide" across timers, unchecked fread in IsZip, unterminated
  strncpy of the ROM filename, and uninitialized i in unzip getShort/getLong.

Warnings, ~200 to zero:
- Wrap commented-out macro blocks in c6502mak.h that spliced lines via
  trailing backslashes in // comments.
- Guard MSVC inlining pragmas behind _MSC_VER; drop C++17-invalid register.
- Move sdlemu_filter.h's global definitions into the .cpp so including TUs
  stop getting private unused copies.
- Remove dead locals; const-correct CSystem/CRom/ContextSave/ContextLoad.

Housekeeping:
- Drop the unused handy-0.90 sourcebase and Makefile-0.90; rename
  Makefile-0.95 to Makefile and update INSTALL.
- Add .gitignore for obj/, the binary, and the copyrighted boot ROM.

+++ src/handy-0.95/c65c02.h
@@ -198,7 +198,7 @@ class C65C02
 		inline bool ContextSave(FILE *fp)
 		{	
 			TRACE_CPU0("ContextSave()");
-			int mPS;
+			ULONG mPS;
 			mPS=PS();
 			if(!fprintf(fp,"C6502::ContextSave")) return 0;
 			if(!fwrite(&mA,sizeof(ULONG),1,fp)) return 0;
@@ -214,7 +214,7 @@ class C65C02
 		inline bool ContextLoad(LSS_FILE *fp)
 		{
 			TRACE_CPU0("ContextLoad()");
-			int mPS;
+			ULONG mPS;
 			char teststr[100]="XXXXXXXXXXXXXXXXXX";
 			if(!lss_read(teststr,sizeof(char),18,fp)) return 0;
 			if(strcmp(teststr,"C6502::ContextSave")!=0) return 0;
